+++ src/org/freeforums/geforce/geffy/commands/CommandGiveCookie.java
@@ -10,15 +10,15 @@ public class CommandGiveCookie implements ICommand<MessageEvent, PrivateMessageE
 
 	public void exeChan(MessageEvent event, String[] args) throws Exception {
 		if(args.length == 0){
-			event.respond("Here's a " + Reference.cookieTypes[Utils.getRandomNumber(Reference.cookieTypes.length - 1)] + " cookie for you!"); 
+			event.respond("Here's a " + Reference.cookieTypes[Utils.getRandomNumber(Reference.cookieTypes.length, true)] + " cookie for you!"); 
 		}else if(args.length == 1){
-			Utils.sendMessage(event, args[0] + ", here's a " + Reference.cookieTypes[Utils.getRandomNumber(Reference.cookieTypes.length - 1)] + " cookie for you!");
+			Utils.sendMessage(event, args[0] + ", here's a " + Reference.cookieTypes[Utils.getRandomNumber(Reference.cookieTypes.length, true)] + " cookie for you!");
 		}
 	}
 
 	public void exePrivate(PrivateMessageEvent event, String[] args) throws Exception {
 		if(args.length == 0){
-			event.respond("Here's a " + Reference.cookieTypes[Utils.getRandomNumber(Reference.cookieTypes.length - 1)] + " cookie for you!"); 
+			event.respond("Here's a " + Reference.cookieTypes[Utils.getRandomNumber(Reference.cookieTypes.length, true)] + " cookie for you!"); 
 		}
 	}