+++ src/org/freeforums/geforce/geffy/main/Reference.java
@@ -1,15 +1,18 @@
 package org.freeforums.geforce.geffy.main;
 
-import org.freeforums.geforce.geffy.commands.*;
+import org.freeforums.geforce.geffy.commands.CommandGiveCookie;
+import org.freeforums.geforce.geffy.commands.CommandHelp;
+import org.freeforums.geforce.geffy.commands.CommandRate;
+import org.freeforums.geforce.geffy.commands.ICommand;
 
 public class Reference {
 	
 	//Important variables and arrays.
-	public static final String commandPrefix = "#";
+	public static final String commandPrefix = "~";
 	
-	public static ICommand[] commands = new ICommand[]{new CommandHelp(), new CommandRank(), new CommandGiveCookie()};
+	public static ICommand[] commands = new ICommand[]{new CommandHelp(), new CommandRate(), new CommandGiveCookie()};
     public static String[] joinedChannels = new String[]{"#Geforce", "#GeforceMods", "#bl4ckscor3", "#Razz"};	
-    public static String[] ignoredUsers = new String[]{"GeffyBot", "bl4ckb0t", "Maunz", "Corobot", "Cadbury"};	
+    public static String[] ignoredUsers = new String[]{"GeffyBot", "Geffy"};	
     
     //Other things.
 	public static String[] cookieTypes = new String[]{"chocolate chip", "peanut butter", "raisin", "oatmeal", "white fudge", "sugar", "frosted", "Oreo"};